Move generated folder
